# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Inventory-App",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api"
        }
    ],
    "paths": {
...
        "/book": {
            "get": {
                "tags": [
                    "book"
                ],
                "summary": "List Books",
                "description": "Get a list of all books.",
                "operationId": "bookList",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Book"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Authorization token invalid."
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "book"
                ],
                "summary": "Create a Book",
                "description": "Create a new book.",
                "operationId": "bookCreate",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "The name of the new inventory item the book belongs to.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "The user who owns the object.",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "location_id",
                        "in": "query",
                        "description": "The location where the book is stored.",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "isbn",
                        "in": "query",
                        "description": "The isbn of the book.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "author",
                        "in": "query",
                        "description": "The author of the book.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "excerpt",
                        "in": "query",
                        "description": "A short text describing the book.",
                        "required": false,
                        "schema": {
                            "type": "text"
                        }
                    },
                    {
                        "name": "release_date",
                        "in": "query",
                        "description": "The date the book was released on.",
                        "required": false,
                        "schema": {
                            "type": "date"
                        },
                        "example": "2021-04-30"
                    },
                    {
                        "name": "language",
                        "in": "query",
                        "description": "The language the book is written in.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Book"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Authorization token invalid."
                    },
                    "422": {
                        "description": "The entered parameters are not valid."
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/book/{id}": {
            "get": {
                "tags": [
                    "book"
                ],
                "summary": "Show Book",
                "description": "Get a specific book.",
                "operationId": "bookShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Book"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Authorization token invalid."
                    },
                    "404": {
                        "description": "The Book does not exist."
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "book"
                ],
                "summary": "Delete Book",
                "description": "Delete a book.",
                "operationId": "bookDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Authorization token invalid."
                    },
                    "404": {
                        "description": "The Book does not exist."
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "book"
                ],
                "summary": "Update a book",
                "description": "Update a existing book.",
                "operationId": "bookUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "isbn",
                        "in": "query",
                        "description": "The isbn of the book.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "author",
                        "in": "query",
                        "description": "The author of the book.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "excerpt",
                        "in": "query",
                        "description": "A short text describing the book.",
                        "required": false,
                        "schema": {
                            "type": "text"
                        }
                    },
                    {
                        "name": "release_date",
                        "in": "query",
                        "description": "The date the book was released on.",
                        "required": false,
                        "schema": {
                            "type": "date"
                        },
                        "example": "2021-04-30"
                    },
                    {
                        "name": "language",
                        "in": "query",
                        "description": "The language the book is written in.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Book"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Authorization token invalid."
                    },
                    "422": {
                        "description": "The entered parameters are not valid."
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Book": {
                "required": [
                    "author",
                    "language"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "isbn": {
                        "type": "string",
                        "example": "ISBN 978-3-86680-192-9"
                    },
                    "author": {
                        "type": "string",
                        "example": "Beatrix Münzer-Glas"
                    },
                    "excerpt": {
                        "type": "text",
                        "example": "Dieser Bildband nimmt den Leser mit auf eine Entdeckungsreise ins vergangene Jahrhundert und präsentiert Liebenswertes und Nachdenkliches aus alten Tagen der Stadt Hof und ihres Umlands. Arnd Kluge und Beatrix Münzer-Glas stellen die Menschen aus der Hofer Region in den Mittelpunkt. Über 160 Aufnahmen lassen die Zeit vom Ende des 19. Jahrhunderts bis 1960 wieder lebendig werden. Eine liebevolle Einladung zum Erinnern, Neu- und Wiederentdecken."
                    },
                    "release_date": {
                        "type": "date",
                        "example": "02-20-2019"
                    },
                    "language": {
                        "type": "string",
                        "example": "Deutsch"
                    },
                    "inventoryItem": {
                        "$ref": "#/components/schemas/InventoryItem"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Book"
                }
            },
            "InventoryItem": {
                "required": [
                    "name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "example": "Buch: Stadt und Landkreis Hof"
                    },
                    "type": {
                        "$ref": "#/components/schemas/Book"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "location": {
                        "$ref": "#/components/schemas/Location"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "InventoryItem"
                }
            },
            "Location": {
                "required": [
                    "branch",
                    "room"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "branch": {
                        "type": "string",
                        "example": "Academy"
                    },
                    "room": {
                        "type": "string",
                        "example": "301"
                    },
                    "shelf": {
                        "type": "string",
                        "example": "1"
                    },
                    "compartment": {
                        "type": "string",
                        "example": "1a"
                    },
                    "description": {
                        "type": "text",
                        "example": "Dieser Ort gehört zu dem Lernenden Jeremy Becker"
                    },
                    "inventoryItems": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/InventoryItem"
                        }
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Location"
                }
            },
            "User": {
                "required": [
                    "name",
                    "email",
                    "password"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "example": "Jeremy Becker"
                    },
                    "email": {
                        "type": "string",
                        "example": "jeremy.becker@twofold.swiss"
                    },
                    "password": {
                        "type": "string"
                    },
                    "location": {
                        "$ref": "#/components/schemas/Location"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Location"
                }
            }
        },
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}